Removed the files that allowed me to avoid seeing so much red
